Refactor everything to JWT
