+++ app/controllers/api/v1/basic_auth_api/protocol_subscriptions_controller.rb
@@ -2,20 +2,24 @@
 
 module Api
   module V1
-    module ApiToken
-      class ProtocolSubscriptionsController < ApiTokenApiController
+    module BasicAuthApi
+      class ProtocolSubscriptionsController < BasicAuthApiController
         before_action :set_person, only: %i[create]
 
         def create
           result = SubscribeToProtocol.run!(
             protocol_name: protocol_subscription_create_params[:protocol_name],
             person: @person,
-            start_date: protocol_subscription_create_params[:start_date] || Time.zone.now
+            start_date: start_date
           )
           render json: result
         end
 
         private
+
+        def start_date
+          Time.zone.parse(protocol_subscription_create_params[:start_date]) || Time.zone.now
+        end
 
         def set_person
           @auth_user ||= AuthUser.find_by(auth0_id_string: protocol_subscription_create_params[:auth0_id_string])
