+++ app/controllers/api/v1/jwt_api/jwt_api_controller.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Api
+  module V1
+    module JwtApi
+      class JwtApiController < ApiController
+        include ::Concerns::IsJwtAuthenticated
+      end
+    end
+  end
+end
