+++ config/routes.rb
@@ -49,14 +49,8 @@ Rails.application.routes.draw do
 
   namespace :api do
     namespace :v1 do
-      resources :questionnaire, only: [:show, :create], param: :key
-      resources :response, only: [:show, :index, :create], param: :uuid do
-        collection do
-          get :completed
-        end
-      end
-      resources :auth_user, only: [:create]
 
+      # JWT and Cookie apis
       # TODO in V2 API, rename to people.
       resources :person do
         collection do
@@ -64,21 +58,41 @@ Rails.application.routes.draw do
           put :update
         end
       end
-      resources :statistics, only: [:index]
-      resources :settings, only: [:index]
-      resources :protocol_subscriptions, only: [:create, :show] do
-        collection do
-          get :mine
+
+      # JWT APIs
+      scope module: :jwt_api do
+        resources :questionnaire, only: [:show, :create], param: :key
+        resources :response, only: [:show, :index, :create], param: :uuid do
+          collection do
+            get :completed
+          end
         end
+        resources :auth_user, only: [:create]
+
+        resources :protocol_subscriptions, only: [:create, :show] do
+          collection do
+            get :mine
+          end
+        end
+        resources :protocol, only: [:index]
       end
-      resources :protocol, only: [:index]
+
+      # Basic auth APIs
+      namespace :basic_auth_api do
+        resources :protocol_subscriptions, only: [:create]
+      end
+
+      # Admin APIs
       namespace :admin do
         resources :team, only: [:show], param: :group
         resources :organization, only: [:show], param: :group
       end
-      namespace :api_token do
-        resources :protocol_subscriptions, only: [:create]
-      end
+
+
+
+      # Public APIs
+      resources :statistics, only: [:index]
+      resources :settings, only: [:index]
     end
   end
   match '/',
