+++ projects/sport-data-valley/README.md
@@ -8,4 +8,13 @@ TOKEN_SIGNATURE_ALGORITHM=HS256
 AUTH0_CLIENT_ID=''
 AUTH0_SIGNING_CERTIFICATE='d434d8c85454ea40a82300a8e53386e95434551c063757f9c7f99a4938a15192336d9ca4d476cf1ab5757605948b2a32b22745d9957d198a6625b99e5108da9b'
 AUTH0_AUDIENCE=''
+
+POSTGRES_DATABASE=sport-data-valley
+MONGO_DATABASE=sport-data-valley
+
+HOST_URL=http://localhost:3002
+HOST_DOMAIN=localhost
+INFO_EMAIL=info@sport-data-valley.nl
+
+SITE_LOCATION=http://localhost:3002
 ```
